refactor: updating package structure for EmailApiClient
